trying to deploy the new rejected function

+++ functions/handlers/rejectQuestion.js
@@ -1,7 +1,11 @@
 module.exports = admin => (snap, context) => {
-  
-  return admin.firestore().collection('rejected')
-    .doc(context.params.id)
-    .create(snap.after.data())
-    .then(() => snap.after.ref.delete()); 
+  const beforeUpdate = snap.before.data();
+  const afterUpdate = snap.after.data();
+
+  if (beforeUpdate.rejected !== afterUpdate.rejected && afterUpdate.rejected) {
+    return admin.firestore().collection('rejected')
+      .doc(context.params.id)
+      .create(snap.after.data())
+      .then(() => snap.after.ref.delete()); 
+  }
 }
